Resolved Spring homework 3: was added and configured migration tools Liquibase, all repositories was  refactored from class to interface and extended from JpaReposotory. Deleted useless code.
