--- service/src/main/java/com/dpdev/dao/BaseRepository.java
@@ -1,52 +1,0 @@
-package com.dpdev.dao;
-
-import com.dpdev.entity.BaseEntity;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
-import javax.persistence.criteria.CriteriaQuery;
-import java.io.Serializable;
-import java.util.List;
-import java.util.Optional;
-
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public abstract class BaseRepository<K extends Serializable, E extends BaseEntity<K>> implements Repository<K, E> {
-
-    private final Class<E> clazz;
-
-    @Getter
-    private final EntityManager entityManager;
-
-    @Override
-    public E save(E entity) {
-        entityManager.persist(entity);
-        return entity;
-    }
-
-    @Override
-    public void delete(E entity) {
-        entityManager.remove(entity);
-        entityManager.flush();
-    }
-
-    @Override
-    public void update(E entity) {
-        entityManager.merge(entity);
-    }
-
-    @Override
-    public Optional<E> findById(K id) {
-        return Optional.ofNullable(entityManager.find(clazz, id));
-    }
-
-    @Override
-    public List<E> findAll() {
-        CriteriaQuery<E> criteria = entityManager.getCriteriaBuilder().createQuery(clazz);
-        criteria.from(clazz);
-        return entityManager.createQuery(criteria)
-                .getResultList();
-    }
-}
