+++ service/src/main/java/com/dpdev/dao/OrderProductRepository.java
@@ -1,15 +1,7 @@
 package com.dpdev.dao;
 
 import com.dpdev.entity.OrderProduct;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import org.springframework.stereotype.Repository;
-
-import javax.persistence.EntityManager;
-
-@Repository
-public class OrderProductRepository extends BaseRepository<Long, OrderProduct> {
-
-    public OrderProductRepository(EntityManager entityManager) {
-        super(OrderProduct.class, entityManager);
-    }
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {
 }
