+++ service/src/main/java/com/dpdev/dao/OrderRepository.java
@@ -1,23 +1,11 @@
 package com.dpdev.dao;
 
 import com.dpdev.entity.Orders;
-import org.hibernate.graph.GraphSemantic;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.persistence.EntityManager;
-import java.util.Map;
+public interface OrderRepository extends JpaRepository<Orders, Long> {
 
-@Repository
-public class OrderRepository extends BaseRepository<Long, Orders> {
-
-    public OrderRepository(EntityManager entityManager) {
-        super(Orders.class, entityManager);
-    }
-
-    public Orders findByIdWithOrderedProducts(Long id) {
-        Map<String, Object> properties = Map.of(
-                GraphSemantic.FETCH.getJpaHintName(), getEntityManager().getEntityGraph("WithUserAndOrderedProducts")
-        );
-        return getEntityManager().find(Orders.class, id, properties);
-    }
+    @EntityGraph(attributePaths = {"user", "orderProducts.product"})
+    public Orders findOrdersAndUsersById(Long id);
 }
