+++ service/src/main/java/com/dpdev/dto/UserWithAvgPriceProjection.java
@@ -1,0 +1,7 @@
+package com.dpdev.dto;
+
+public interface UserWithAvgPriceProjection {
+    String getEmail();
+
+    Double getAvgPrice();
+}
