+++ service/src/main/resources/application.yml
@@ -9,7 +9,7 @@ spring:
       hibernate.format_sql: true
     show-sql: true
     hibernate:
-      ddl-auto: create
+      ddl-auto: validate
 
 logging:
   pattern:
