+++ service/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,0 +1,5 @@
+databaseChangeLog:
+  - include:
+      file: db/changelog/db.changelog-1.0.sql
+  - include:
+      file: db/changelog/db.changelog-2.0.sql
