--- service/src/main/resources/log4j.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration>
-
-    <appender name="console" class="org.apache.log4j.ConsoleAppender">
-        <param name="target" value="System.out"/>
-
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="conversionPattern" value="[%d{HH:mm:ss,SSS}] %p [%c: %L] %m%n"/>
-        </layout>
-    </appender>
-
-    <root>
-        <level value="info"/>
-        <appender-ref ref="console"/>
-    </root>
-
-</log4j:configuration>
