+++ service/src/test/java/com/dpdev/integration/IntegrationTestBase.java
@@ -9,10 +9,9 @@ import com.dpdev.entity.enums.Brand;
 import com.dpdev.entity.enums.OrderStatus;
 import com.dpdev.entity.enums.ProductType;
 import com.dpdev.entity.enums.Role;
-import com.dpdev.integration.util.TestDataImporter;
-import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,7 +22,8 @@ import java.time.Instant;
 
 import static com.dpdev.integration.util.ContainerTestUtil.postgres;
 
-@SpringBootTest()
+@SpringBootTest
+@ActiveProfiles("test")
 @Transactional
 public abstract class IntegrationTestBase {
 
@@ -33,13 +33,6 @@ public abstract class IntegrationTestBase {
     @DynamicPropertySource
     static void setProperties(DynamicPropertyRegistry registry) {
         registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-    }
-
-    @BeforeEach
-    void startSession() {
-        TestDataImporter.importData(entityManager);
     }
 
     public static User createUser() {
