+++ service/src/test/resources/application-test.yml
@@ -1,0 +1,8 @@
+spring:
+  datasource:
+    username: test
+    password: test
+  jpa:
+    properties:
+      hibernate:
+        ddl-auto: create
